digital_products_orders/requirements.txt
requests
numpy==1.26.4
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
openpyxl
mysql-connector-python
google-auth
google-api-python-client==2.31.0
oauth2client==4.1.3
ShopifyApi